color, bloom, ui

+++ Project Grievous/Assets/Shaders/DoorShader.shader
@@ -37,10 +37,10 @@ Shader "Custom/DoorShader"
         UNITY_INSTANCING_BUFFER_START(Props)
         UNITY_INSTANCING_BUFFER_END(Props)
 
-        float3 FlowUVW(float2 uv, float2 flowVector, float tiling, float time, bool flowB) {
-            float phaseOffset = flowB ? 0.5 : 0;
-            float progress = frac(time + phaseOffset);
-            float3 uvw;
+        half3 FlowUVW(half2 uv, half2 flowVector, half tiling, half time, bool flowB) {
+            half phaseOffset = flowB ? 0.5 : 0;
+            half progress = frac(time + phaseOffset);
+            half3 uvw;
             uvw.xy = uv - flowVector * progress;
             uvw.xy *= tiling;
             uvw.xy += phaseOffset;
@@ -50,13 +50,13 @@ Shader "Custom/DoorShader"
 
         void surf(Input IN, inout SurfaceOutputStandard o)
         {
-            float2 flowVector = tex2D(_FlowMap, IN.uv_MainTex).rg * 2 - 1;
-            float noise = tex2D(_FlowMap, IN.uv_MainTex).a;
-            float time = _Time.y * _Speed + noise;
-            float3 uvwA = FlowUVW(
+            half2 flowVector = tex2D(_FlowMap, IN.uv_MainTex).rg * 2 - 1;
+            half noise = tex2D(_FlowMap, IN.uv_MainTex).a;
+            half time = _Time.y * _Speed + noise;
+            half3 uvwA = FlowUVW(
                 IN.uv_MainTex, flowVector, _Tiling, time, false
             );
-            float3 uvwB = FlowUVW(
+            half3 uvwB = FlowUVW(
                 IN.uv_MainTex, flowVector, _Tiling, time, true
             );
             fixed4 texA = tex2D(_MainTex, uvwA.xy) * uvwA.z;
